feat(helpers): allow optionally overriding set expires_in
There are times, when using the mixed-in Helpers, that you don't
want the default SEVEN_DAYS dedup expiration.

This update the helpers to add a hook method, dedupe_expires_in,
that can be overridden to use a custom expires_in value. It defaults
to the SEVEN_DAYS default just as before so is backwards compatible.